rework race condition found in issue #1491
See https://github.com/zendesk/maxwell/issues/1491, we need to wait for
mysql-binlog-connector to actually tell us it's disconnected before we
can perform a reconnect.  otherwise we could possibly race it to the end
of its function.
