--- src/main/java/com/zendesk/maxwell/replication/BinlogConnectorLifecycleListener.java
@@ -1,37 +1,0 @@
-package com.zendesk.maxwell.replication;
-
-import com.github.shyiko.mysql.binlog.BinaryLogClient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-class BinlogConnectorLifecycleListener implements BinaryLogClient.LifecycleListener {
-	private static final Logger LOGGER = LoggerFactory.getLogger(BinlogConnectorLifecycleListener.class);
-	private BinlogConnectorReplicator replicator;
-
-	public BinlogConnectorLifecycleListener(BinlogConnectorReplicator replicator) {
-		this.replicator = replicator;
-	}
-
-	@Override
-	public void onConnect(BinaryLogClient client) {
-		LOGGER.info("Binlog connected.");
-	};
-
-	@Override
-	public void onCommunicationFailure(BinaryLogClient client, Exception ex) {
-		LOGGER.warn("Communication failure.", ex);
-		replicator.onCommunicationFailure(ex);
-	}
-
-	@Override
-	public void onEventDeserializationFailure(BinaryLogClient client, Exception ex) {
-		LOGGER.warn("Event deserialization failure.", ex);
-		LOGGER.warn("cause: ", ex.getCause());
-	}
-
-	@Override
-	public void onDisconnect(BinaryLogClient client) {
-		LOGGER.info("Binlog disconnected.");
-	}
-}
-
